pyproject.toml
[project]
name = "ml-cookbook-project"
version = "0.1.0"
description = "Cookbook de prototypage ML avec PyTorch et Lightning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "lightning>=2.2,<3.0",
  "torch>=2.2,<3.0",
  "torchvision>=0.17,<1.0",
  "numpy>=1.26,<3.0",
  "pyyaml>=6.0,<7.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0,<9.0",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ml_cookbook"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
